fix flightsScreen
FlightsScreen view refactored, functions separated
